refactor: 이벤트 리스너를 별도로 분리
