--- src/main/java/umc/cockple/demo/domain/chat/service/listener/ChatEventListener.java
@@ -1,191 +1,0 @@
-package umc.cockple.demo.domain.chat.service.listener;
-
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.context.event.EventListener;
-import org.springframework.scheduling.annotation.Async;
-import org.springframework.stereotype.Component;
-import org.springframework.transaction.event.TransactionPhase;
-import org.springframework.transaction.event.TransactionalEventListener;
-import umc.cockple.demo.domain.chat.dto.WebSocketMessageDTO;
-import umc.cockple.demo.domain.chat.dto.WebSocketMessageDTO.ChatRoomListUpdate.LastMessageUpdate;
-import umc.cockple.demo.domain.chat.enums.WebSocketMessageType;
-import umc.cockple.demo.domain.chat.events.ChatListSubscriptionEvent;
-import umc.cockple.demo.domain.chat.events.ChatMessageSendEvent;
-import umc.cockple.demo.domain.chat.events.ChatRoomListUpdateEvent;
-import umc.cockple.demo.domain.chat.events.ChatRoomSubscriptionEvent;
-import umc.cockple.demo.domain.chat.events.ChatUnreadStatusUpdateEvent;
-import umc.cockple.demo.domain.chat.service.ChatUnreadQueryService;
-import umc.cockple.demo.domain.chat.repository.redis.ChatListSubscriptionStore;
-import umc.cockple.demo.domain.chat.service.websocket.ChatRoomListCacheService;
-import umc.cockple.demo.domain.chat.service.websocket.broadcast.ChatRoomListUpdateBroadcaster;
-import umc.cockple.demo.domain.chat.service.websocket.broadcast.ChatRoomListUpdateData;
-import umc.cockple.demo.domain.chat.service.websocket.send.ChatSendService;
-import umc.cockple.demo.domain.chat.service.websocket.session.ChatMessageSender;
-import umc.cockple.demo.domain.chat.service.websocket.subscription.ChatRoomSubscriptionService;
-import umc.cockple.demo.domain.notification.events.ChatNotificationEvent;
-import umc.cockple.demo.domain.notification.service.ChatPushNotificationService;
-import umc.cockple.demo.domain.party.events.PartyMemberJoinedEvent;
-
-import java.time.LocalDateTime;
-import java.util.HashMap;
-import java.util.Map;
-
-@Component
-@RequiredArgsConstructor
-@Slf4j
-public class ChatEventListener {
-
-    private final ChatSendService chatSendService;
-    private final ChatRoomSubscriptionService chatRoomSubscriptionService;
-    private final ChatRoomListCacheService chatRoomListCacheService;
-    private final ChatRoomListUpdateBroadcaster chatRoomListUpdateBroadcaster;
-    private final ChatListSubscriptionStore chatListSubscriptionStore;
-    private final ChatMessageSender chatMessageSender;
-    private final ChatPushNotificationService chatPushNotificationService;
-    private final ChatUnreadQueryService chatUnreadQueryService;
-
-
-    @EventListener
-    @Async
-    public void handleChatMessageSend(ChatMessageSendEvent event) {
-        log.info("메시지 전송 이벤트 처리 - 채팅방: {}, 발신자: {}",
-                event.chatRoomId(), event.senderId());
-        try {
-            chatSendService
-                    .sendMessage(event.chatRoomId(), event.content(), event.files(), event.senderId());
-        } catch (Exception e) {
-            log.error("메시지 전송 이벤트 처리 중 오류 발생", e);
-        }
-    }
-
-    @TransactionalEventListener(phase = TransactionPhase.AFTER_COMMIT)
-    @Async
-    public void handleChatNotification(ChatNotificationEvent event) {
-        log.info("채팅 알림 이벤트 처리 - 채팅방: {}, 발신자: {}", event.chatRoomId(), event.senderId());
-        try {
-            chatPushNotificationService.sendPush(event);
-        } catch (Exception e) {
-            log.error("채팅 알림 이벤트 처리 중 오류 발생 - 채팅방: {}", event.chatRoomId(), e);
-        }
-    }
-
-    @TransactionalEventListener(phase = TransactionPhase.AFTER_COMMIT) //트랜잭션이 커밋된 후에 실행
-    @Async
-    public void handlePartyMemberChanged(PartyMemberJoinedEvent event) {
-        switch (event.action()) {
-            case JOINED -> chatSendService.sendSystemMessage(event.partyId(),
-                    event.memberName() + "님이 모임에 참여하셨습니다.");
-            case LEFT -> chatSendService.sendSystemMessage(event.partyId(),
-                    event.memberName() + "님이 모임을 떠나셨습니다.");
-        }
-    }
-
-    @EventListener
-    public void handleChatRoomSubscription(ChatRoomSubscriptionEvent event) {
-        log.info("채팅방 구독 이벤트 처리 - 채팅방: {}, 사용자: {}, 액션: {}",
-                event.chatRoomId(), event.memberId(), event.action());
-
-        try {
-            switch (event.action()) {
-                case "SUBSCRIBE" -> {
-                    chatRoomSubscriptionService.subscribeToChatRoom(event.chatRoomId(), event.memberId());
-                    log.info("사용자 {}가 채팅방 {}를 구독했습니다.", event.memberId(), event.chatRoomId());
-                }
-                case "UNSUBSCRIBE" -> {
-                    chatRoomSubscriptionService.unsubscribeToChatRoom(event.chatRoomId(), event.memberId());
-                    log.info("사용자 {}가 채팅방 {}를 구독해제했습니다.", event.memberId(), event.chatRoomId());
-                }
-                default -> log.warn("알 수 없는 구독 액션: {}", event.action());
-            }
-        } catch (Exception e) {
-            log.error("채팅방 구독 이벤트 처리 중 오류 발생", e);
-        }
-    }
-
-    @EventListener
-    @Async
-    public void handleChatListSubscription(ChatListSubscriptionEvent event) {
-        log.info("채팅방 목록 구독 이벤트 처리 시작 - 멤버: {}, 액션: {}, 채팅방 수: {}",
-                event.memberId(), event.action(), event.chatRoomIds().size());
-
-        try {
-            switch (event.action()) {
-                case "SUBSCRIBE" -> {
-                    chatListSubscriptionStore.subscribeToChatList(event.memberId(), event.chatRoomIds());
-                    log.info("채팅방 목록 구독 완료 - 멤버: {}, 채팅방 수: {}", event.memberId(), event.chatRoomIds().size());
-                }
-                case "UNSUBSCRIBE" -> {
-                    chatListSubscriptionStore.unsubscribeFromChatList(event.memberId(), event.chatRoomIds());
-                    log.info("채팅방 목록 구독 해제 완료 - 멤버: {}, 채팅방 수: {}", event.memberId(), event.chatRoomIds().size());
-                }
-                default -> log.warn("알 수 없는 채팅방 목록 구독 액션: {}", event.action());
-            }
-
-        } catch (Exception e) {
-            log.error("채팅방 목록 구독 이벤트 처리 중 오류 발생 - 멤버: {}, 액션: {}",
-                    event.memberId(), event.action(), e);
-        }
-    }
-
-    @TransactionalEventListener(phase = TransactionPhase.AFTER_COMMIT)
-    @Async
-    public void handleChatRoomListUpdate(ChatRoomListUpdateEvent event) {
-        log.info("채팅방 목록 업데이트 이벤트 처리 시작 - 채팅방: {}", event.chatRoomId());
-
-        try {
-            chatRoomListCacheService.evictLastMessage(event.chatRoomId());
-
-            Map<Long, ChatRoomListUpdateData> memberUpdateData = new HashMap<>();
-
-            LastMessageUpdate lastMessageUpdate = LastMessageUpdate.builder()
-                    .content(event.content())
-                    .timestamp(event.timestamp())
-                    .messageType(event.messageType())
-                    .build();
-
-            for (Map.Entry<Long, Integer> entry : event.memberUnreadCounts().entrySet()) {
-                Long memberId = entry.getKey();
-                Integer unreadCount = entry.getValue();
-
-                memberUpdateData.put(memberId, ChatRoomListUpdateData.builder()
-                        .lastMessage(lastMessageUpdate)
-                        .unreadCount(unreadCount)
-                        .build());
-            }
-
-            chatRoomListUpdateBroadcaster.broadcast(event.chatRoomId(), memberUpdateData);
-
-            log.info("채팅방 목록 업데이트 이벤트 처리 완료 - 채팅방: {}", event.chatRoomId());
-
-        } catch (Exception e) {
-            log.error("채팅방 목록 업데이트 이벤트 처리 중 오류 발생 - 채팅방: {}", event.chatRoomId(), e);
-        }
-    }
-
-    @TransactionalEventListener(phase = TransactionPhase.AFTER_COMMIT)
-    @Async
-    public void handleChatUnreadStatusUpdate(ChatUnreadStatusUpdateEvent event) {
-        log.info("채팅 안읽음 상태 업데이트 이벤트 처리 시작 - 대상자: {}명", event.targetMemberIds().size());
-
-        for (Long memberId : event.targetMemberIds()) {
-            try {
-                boolean hasPartyUnread = chatUnreadQueryService.hasPartyUnreadMessages(memberId);
-                boolean hasDirectUnread = chatUnreadQueryService.hasDirectUnreadMessages(memberId);
-
-                WebSocketMessageDTO.UnreadStatusUpdateMessage message =
-                        WebSocketMessageDTO.UnreadStatusUpdateMessage.builder()
-                                .type(WebSocketMessageType.UNREAD_STATUS_UPDATE)
-                                .hasUnread(hasPartyUnread || hasDirectUnread)
-                                .hasPartyUnread(hasPartyUnread)
-                                .hasDirectUnread(hasDirectUnread)
-                                .timestamp(LocalDateTime.now())
-                                .build();
-
-                chatMessageSender.send(memberId, message);
-            } catch (Exception e) {
-                log.error("채팅 안읽음 상태 업데이트 처리 실패 - 멤버: {}", memberId, e);
-            }
-        }
-    }
-}
